+++ jstests/core/plan_cache_clear.js
@@ -14,6 +14,14 @@
 // ]
 
 (function() {
+'use strict';
+load("jstests/libs/sbe_util.js");  // For checkSBEEnabled.
+
+if (checkSBEEnabled(db, ["featureFlagSbePlanCache"])) {
+    jsTest.log("Skipping test because SBE and SBE plan cache are both enabled.");
+    return;
+}
+
 const coll = db.jstests_plan_cache_clear;
 coll.drop();
