+++ jstests/core/plan_cache_list_plans.js
@@ -16,6 +16,13 @@
 
 (function() {
 "use strict";
+load("jstests/libs/sbe_util.js");  // For checkSBEEnabled.
+
+if (checkSBEEnabled(db, ["featureFlagSbePlanCache"])) {
+    jsTest.log("Skipping test because SBE and SBE plan cache are both enabled.");
+    return;
+}
+
 let coll = db.jstests_plan_cache_list_plans;
 coll.drop();
