+++ jstests/core/plan_cache_list_shapes.js
@@ -13,6 +13,14 @@
 //   assumes_unsharded_collection,
 // ]
 (function() {
+'use strict';
+load("jstests/libs/sbe_util.js");  // For checkSBEEnabled.
+
+if (checkSBEEnabled(db, ["featureFlagSbePlanCache"])) {
+    jsTest.log("Skipping test because SBE and SBE plan cache are both enabled.");
+    return;
+}
+
 const coll = db.jstests_plan_cache_list_shapes;
 coll.drop();
