+++ jstests/core/plan_cache_sbe.js
@@ -19,6 +19,11 @@
 
 load("jstests/libs/sbe_util.js");  // For checkSBEEnabled.
 
+if (checkSBEEnabled(db, ["featureFlagSbePlanCache"])) {
+    jsTest.log("Skipping test because SBE and SBE plan cache are both enabled.");
+    return;
+}
+
 const coll = db.plan_cache_sbe;
 coll.drop();
