+++ jstests/core/wildcard_index_cached_plans.js
@@ -22,6 +22,11 @@ load("jstests/libs/collection_drop_recreate.js");  // For assert[Drop|Create]Col
 load('jstests/libs/fixture_helpers.js');  // For getPrimaryForNodeHostingDatabase and isMongos.
 load("jstests/libs/sbe_util.js");         // For checkSBEEnabled.
 
+if (checkSBEEnabled(db, ["featureFlagSbePlanCache"])) {
+    jsTest.log("Skipping test because SBE and SBE plan cache are both enabled.");
+    return;
+}
+
 const coll = db.wildcard_cached_plans;
 coll.drop();
